Consolidate into one GLTF importer
