+++ addons/io_hubs_addon/components/definitions/visible.py
@@ -3,6 +3,7 @@ from bpy.props import BoolProperty
 from ..hubs_component import HubsComponent
 from ..types import Category, PanelType, NodeType
 
+
 class Visible(HubsComponent):
     _definition = {
         'name': 'visible',
